Move CanQueryChainStatusWithHeight to sovereign-rollup-components
