+++ crates/sovereign/sovereign-rollup-components/src/traits/mod.rs
@@ -1,1 +1,2 @@
+pub mod chain_status;
 pub mod json_rpc_client;
